// rickApp/src/app/interfaces/location.interface.ts
// Generated by https://quicktype.io

export interface Location {
  info:    Info;
  results: Result[];
}

export interface Info {
  count: number;
  pages: number;
  next:  string;
  prev:  null;
}

export interface Result {
  id:        number;
  name:      string;
  type:      string;
  dimension: string;
  residents: string[];
  url:       string;
  created:   string;
  showResidents?: boolean;
  characters?: Character[];
  showCharacters?: boolean;
}

export interface Character {
  id:       number;
  name:     string;
  type:     string;
  origin:   Location;
  location: Location;
  image:    string;
  episode:  string[];
  url:      string;
  created:  string;
  species:  Species;
  gender:   Gender;
  status:   Status;
}
export enum Species {
  Alien = "Alien",
  Human = "Human",
}

export enum Gender {
  Female = "Female",
  Male = "Male",
  Unknown = "unknown",
}
export enum Status {
  Alive = "Alive",
  Dead = "Dead",
  Unknown = "unknown",
}